+++ src/main/java/com/junwoo/ott/domain/payment/dto/remote/PaymentDto.java
@@ -1,17 +1,17 @@
-package com.junwoo.ott.domain.payment.dto.body;
+package com.junwoo.ott.domain.payment.dto.remote;
 
-import com.junwoo.ott.domain.payment.dto.body.payment.CacheReceiptDto;
-import com.junwoo.ott.domain.payment.dto.body.payment.CancelDto;
-import com.junwoo.ott.domain.payment.dto.body.payment.CardDto;
-import com.junwoo.ott.domain.payment.dto.body.payment.CheckoutDto;
-import com.junwoo.ott.domain.payment.dto.body.payment.DiscountDto;
-import com.junwoo.ott.domain.payment.dto.body.payment.EasyPayDto;
-import com.junwoo.ott.domain.payment.dto.body.payment.FailureDto;
-import com.junwoo.ott.domain.payment.dto.body.payment.GiftCertificateDto;
-import com.junwoo.ott.domain.payment.dto.body.payment.MobilePhoneDto;
-import com.junwoo.ott.domain.payment.dto.body.payment.ReceiptDto;
-import com.junwoo.ott.domain.payment.dto.body.payment.TransferDto;
-import com.junwoo.ott.domain.payment.dto.body.payment.VirtualAccountDto;
+import com.junwoo.ott.domain.payment.dto.remote.payment.CacheReceiptDto;
+import com.junwoo.ott.domain.payment.dto.remote.payment.CancelDto;
+import com.junwoo.ott.domain.payment.dto.remote.payment.CardDto;
+import com.junwoo.ott.domain.payment.dto.remote.payment.CheckoutDto;
+import com.junwoo.ott.domain.payment.dto.remote.payment.DiscountDto;
+import com.junwoo.ott.domain.payment.dto.remote.payment.EasyPayDto;
+import com.junwoo.ott.domain.payment.dto.remote.payment.FailureDto;
+import com.junwoo.ott.domain.payment.dto.remote.payment.GiftCertificateDto;
+import com.junwoo.ott.domain.payment.dto.remote.payment.MobilePhoneDto;
+import com.junwoo.ott.domain.payment.dto.remote.payment.ReceiptDto;
+import com.junwoo.ott.domain.payment.dto.remote.payment.TransferDto;
+import com.junwoo.ott.domain.payment.dto.remote.payment.VirtualAccountDto;
 import com.junwoo.ott.global.customenum.payment.PaymentStatusType;
 import com.junwoo.ott.global.customenum.payment.PaymentType;
 import com.junwoo.ott.global.customenum.payment.common.PaymentMethodType;
