+++ src/main/java/com/junwoo/ott/domain/payment/dto/remote/payment/FailureDto.java
@@ -1,4 +1,4 @@
-package com.junwoo.ott.domain.payment.dto.body.payment;
+package com.junwoo.ott.domain.payment.dto.remote.payment;
 
 
 import lombok.AllArgsConstructor;
