+++ src/main/java/com/junwoo/ott/domain/payment/dto/remote/payment/GiftCertificateDto.java
@@ -1,19 +1,18 @@
-package com.junwoo.ott.domain.payment.dto.body.payment;
+package com.junwoo.ott.domain.payment.dto.remote.payment;
 
 import com.junwoo.ott.global.customenum.payment.common.SettlementStatusType;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
-
 /**
- * 계좌이체로 결제했을 때 이체 정보가 담기는 객체입니다.
+ * 상품권으로 결제하면 제공되는 상품권 결제 관련 정보입니다.
  */
 @Getter
 @AllArgsConstructor
-public class TransferDto {
+public class GiftCertificateDto {
 
-  // 은행 숫자 코드입니다. 은행 코드와 증권사 코드를 참고하세요.
-  private String bankCode;
+  // 결제 승인번호입니다. 최대 길이는 8자입니다.
+  private String approveNo;
   // 정산 상태입니다. 정산이 아직 되지 않았다면 INCOMPLETED, 정산이 완료됐다면 COMPLETED 값이 들어옵니다.
   private SettlementStatusType settlementStatus;
