+++ src/main/java/com/junwoo/ott/domain/payment/dto/remote/payment/ReceiptDto.java
@@ -1,11 +1,11 @@
-package com.junwoo.ott.domain.payment.dto.body.payment;
+package com.junwoo.ott.domain.payment.dto.remote.payment;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public class CheckoutDto {
+public class ReceiptDto {
 
   private String url;
